Fixed MeshComponent Implementation and extra findings
noted:
1. error occurs because we need to multiply a vec3 by a float, which needs to be done explicitly in GLM. method to fix:
Explicitly casting the float _normalLength to double since we're using dvec3 (double precision vectors)
This ensures proper type matching for the multiplication operator

2. extra vec2 defintion needed to be added, plus more colors in namespace Colors

+++ SpaghettiMaker/SpaghettiEngine/MeshComponent.cpp
@@ -1,3 +1,4 @@
+// SpaghettiEngine/Graphics/MeshComponent.cpp
 #include "MeshComponent.h"
 #include "GameObject.h"
 #include "TransformComponent.h"
@@ -94,7 +95,8 @@ void MeshComponent::OnUpdate() {
 
         for (const auto& vertex : _vertices) {
             vec3 start = vertex.position;
-            vec3 end = start + (vertex.normal * _normalLength);
+            // Fixed: Multiply vec3 by scalar using proper GLM syntax
+            vec3 end = start + (vertex.normal * static_cast<double>(_normalLength));
 
             glVertex3dv(glm::value_ptr(start));
             glVertex3dv(glm::value_ptr(end));
